[refactor] Modify date format and file path for Calendar component
