+++ src/components/myPage/Diary/DiaryDetails.tsx
